Further abstraction away from bot config files to GenericBot
I think I've abstracted as much as I reasonably can here. Moved the functions which get the bot secrets to the GenericBot class itself, as well as the creation and maintenance of the CommandLogger object and other code common to both bots. Now all that is required is to set up a GenericBot.Command.Collection object, create a new GenericBot object using the Command.Collection, then with the newly created bot run the configure() and login() methods.

Modified the README to document these changes and provide a sample bot config file.

Resolves #10